Actually working
all functions send values to respective values or streams.
fixed the print(TEST MORE) longs on the fringe were causeing problems with x0fff.... it should be xfffff....
TODO: need to clean up
need to comment(still)
need to test the asserts for unser. and for ser.(string)
need to remove unecessary code.